Rename components to better reflect the new toggle component
