+++ src/panels/NightSkyPanel/tabs/MarkingsTab/SceneGraphNodeToggle.tsx
@@ -2,7 +2,7 @@ import { usePropertyOwner, usePropertyOwnerVisibility } from '@/hooks/propertyOw
 import { Identifier } from '@/types/types';
 import { sgnRenderableUri, sgnUri } from '@/util/propertyTreeHelpers';
 
-import { MarkingBoxLayout } from './MarkingBoxLayout';
+import { MarkingsToggleLayout } from './MarkingsToggleLayout';
 
 interface Props {
   title: string;
@@ -13,14 +13,14 @@ interface Props {
   identifier: Identifier;
 }
 
-export function SceneGraphNodeBox({ title, icon, identifier }: Props) {
+export function SceneGraphNodeToggle({ title, icon, identifier }: Props) {
   const uri = sgnRenderableUri(sgnUri(identifier));
   const propertyOwner = usePropertyOwner(uri);
 
   const { isVisible, setVisibility } = usePropertyOwnerVisibility(uri);
 
   return (
-    <MarkingBoxLayout
+    <MarkingsToggleLayout
       onClick={() => setVisibility(!isVisible)}
       checked={isVisible || false}
       aria-label={title} // TODO: Update with correct aria-label
